Expand logic around persistence and reloading
We want to allow users to link Consul clusters that already exist to
HCP. Existing clusters need care when bootstrapped by HCP, since we do
not want to do things like change ACL/TLS settings for a running
cluster.

The plan is to only send down a management token to existing clusters
and nothing else. This means that all the logic around persisting and
reloading TLS and JSON config gets skipped in this path.

This commit also adds some light validation when loading data from disk,
to raise errors when we should re-fetch data from HCP.