Remove inference about existing clusters
Rather than having separate code paths for new and existing clusters,
this commit consolidates persist/load logic so that we validate and
persist everything that is received.

When restarting, we check for completeness based on an empty file called
"success". If present, we expect files on disk to be complete for this
cluster type and proceed to load all the ones that are available.