Ensure the initial_management is sanitized
When the ACL system can be bootstrapped with the
CloudConfig.ManagementToken, there is no need to ever use the
initial_management token for HCP's purposes.

However, older versions of Consul that can be bootstrapped from CCM
still require an initial_management token. For now CCM does not have an
awareness of what version of Consul requested bootstrap configuration.

To account for that, we're going to temporarily sanitize it from the
received config.